commit_fix_merge_issue

+++ Supply/Dto/RecommendingSupplyDto.cs
@@ -4,7 +4,7 @@
     {
         public double amount { get; set; }
         public Supply.Supply fertilizer { get; set; }
-        public RecommendingSupplyDto(long amount, Supply.Supply fertilizer)
+        public RecommendingSupplyDto(double amount, Supply.Supply fertilizer)
         {
             this.amount = amount;
             this.fertilizer = fertilizer;
